+++ .env.keystore.example
@@ -1,0 +1,8 @@
+DEV_DEPLOYER_PRIVATE_KEY_PATH="path/to/deployer.txt"
+DEV_DIRECTORY_DEPLOYER_PRIVATE_KEY_PATH="path/to/directory-deployer.txt"
+
+STAGING_DEPLOYER_PRIVATE_KEY_PATH="path/to/deployer.txt"
+STAGING_DIRECTORY_DEPLOYER_PRIVATE_KEY_PATH="path/to/directory-deployer.txt"
+
+PROD_DEPLOYER_PRIVATE_KEY_PATH="path/to/deployer.txt"
+PROD_DIRECTORY_DEPLOYER_PRIVATE_KEY_PATH="path/to/directory-deployer.txt"
