+++ .env.staging
@@ -8,5 +8,3 @@ ADMIN_ORACLE="0xD2bb9AE090422beaE33c6D7655de1E5d390AB8C6"
 SANCTIONS_LIST_ADDRESS="0x0574C4ee6AfAEe9EcA287B30Ef1937E82e5e95a9"
 WETH_ADDRESS="0xEA24abFF2D781d3c2dac394c46686b5E43497B6a"
 CHAIN_ID="17000"
-DEPLOYER_PRIVATE_KEY_PATH="/Users/Keys/Nodeset/Constellation/Staging/deployer.txt"
-DIRECTORY_DEPLOYER_PRIVATE_KEY_PATH="/Users/Keys/Nodeset/Constellation/Staging/directory-deployer.txt"
