+++ .gitignore
@@ -19,3 +19,4 @@ artifacts
 
 #OpenZeppelin
 .openzeppelin/unknown-*.json
+*.keystore
